+++ output/then-ajax.d.ts
@@ -1,6 +1,8 @@
-import ajax, { IAjaxOptions } from './ajax';
-declare function thenAjax(op: IAjaxOptions): Promise<any>;
-declare function thenGet(url: string, data?: any, options?: IAjaxOptions): Promise<any>;
-declare function thenPost(url: string, data?: any, options?: IAjaxOptions): Promise<any>;
-declare function thenJsonp(url: string, data?: any, options?: IAjaxOptions): Promise<any>;
-export { thenAjax, ajax, IAjaxOptions, thenGet, thenPost, thenJsonp };
+import { AjaxOptions as _AjaxOptions, ajax as _ajax } from './ajax';
+export declare const ajax: typeof _ajax;
+export interface AjaxOptions extends _AjaxOptions {
+}
+export declare function thenAjax(op: AjaxOptions): Promise<any>;
+export declare function thenGet(url: string, data?: any, options?: AjaxOptions): Promise<any>;
+export declare function thenPost(url: string, data?: any, options?: AjaxOptions): Promise<any>;
+export declare function thenJsonp(url: string, data?: any, options?: AjaxOptions): Promise<any>;