+++ history.md
@@ -1,9 +1,12 @@
 # 历史版本
+## 0.2.0(2020-01-11)
+* feat: 调整格式让组件支持 tree-shaking
+
 ## 0.1.2(2019-07-28)
-* [ADD] 添加 `timeout` 自动 抛出 new Error('timeout') 处理
+* feat: 添加 `timeout` 自动 抛出 new Error('timeout') 处理
 
 ## 0.1.1(2019-07-01)
-* [DEL] 去掉 `export default` 以兼容 ie8
+* del: 去掉 `export default` 以兼容 ie8
 
 ## 0.1.0(2019-07-01)
-* [ADD] 诞生
+* feat: 诞生